+++ dynamicarray.h
@@ -6,13 +6,13 @@
 #define THREADPROJEKT_DYNAMICARRAY_H
 typedef struct {
     char **array;
-    size_t used;
-    size_t size;
-    size_t strsize;
+    unsigned long int used;
+    unsigned long int size;
+    unsigned long int strsize;
 } Array;
 
 
-void initArray(Array *a, size_t initialSize, size_t stringsize);
+void initArray(Array *a, unsigned long int initialSize, unsigned long int stringsize);
 void addString(Array *a, char *element);
 int contains(Array* a, char * element);
 void freeArray(Array *a);
